refactor: theme slice
