+++ src/components/AccountAdminNavbar/AdminAppNewAuthorNavbar/AdminAppNewAuthorNavbar.jsx
@@ -6,7 +6,8 @@ import rejectIconB from 'images/account/cancel-icon_black.svg';
 import backIconW from 'images/account/logout-icon.svg';
 import backIconB from 'images/account/logout-icon_black.svg';
 import { useSelector } from 'react-redux';
-import { getCurrentTheme, getDecision } from 'store/selectors';
+import { getDecision } from 'store/selectors';
+import { getCurrentTheme } from 'store/slices/theme';
 import { useLocation, useNavigate } from 'react-router-dom';
 import adminApi from 'utils/api/admin';
 import { Modal } from 'components';
