+++ src/components/AccountAdminNavbar/AdminNavbar/AdminNavbar.jsx
@@ -9,7 +9,7 @@ import logoutIconW from 'images/account/logout-icon.svg';
 import logoutIconB from 'images/account/logout-icon_black.svg';
 import HashButton from 'components/LinkButton/HashButton';
 import { useSelector } from 'react-redux';
-import { getCurrentTheme } from 'store/selectors';
+import { getCurrentTheme } from 'store/slices/theme';
 import { adminHash } from 'utils/constants';
 
 export default function AdminNavbar({ logout }) {
