+++ src/components/AccountAuthor/MobileAccountAuthor/MobileAccountAuthor.jsx
@@ -1,7 +1,7 @@
 import './MobileAccountAuthor.css';
 import { useSelector } from 'react-redux';
 
-import { getCurrentTheme } from 'store/selectors';
+import { getCurrentTheme } from 'store/slices/theme';
 import Button from 'components/Button/Button';
 import mobileAccountImgLight from 'images/account/mobile/mobile-account-img-light.svg';
 import mobileAccountImgDark from 'images/account/mobile/mobile-account-img-dark.svg';
