+++ src/components/ArticlesTree/TreeItem/TreeItem.jsx
@@ -3,10 +3,15 @@ import { useNavigate, useParams } from 'react-router-dom';
 import { TreeList } from 'components';
 import './TreeItem.css';
 import { useSelector } from 'react-redux';
-import { getCurrentTheme } from 'store/selectors';
+import { getCurrentTheme } from 'store/slices/theme';
 import findId from './findId';
 
-export default function TreeItem({ title = '', data = null, language, status }) {
+export default function TreeItem({
+  title = '',
+  data = null,
+  language,
+  status,
+}) {
   const navigate = useNavigate();
   const { articleId } = useParams();
   const theme = useSelector(getCurrentTheme);
@@ -14,7 +19,7 @@ export default function TreeItem({ title = '', data = null, language, status }) 
     if (status === false) return false;
     return data ? findId(data, articleId) : false;
   });
-  
+
   const hasChildren = data && typeof data === 'object';
   const id = hasChildren ? data.id : data;
   const isActive = articleId ? articleId === id : false;
