+++ src/components/Author/Author.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { useSelector } from 'react-redux';
 import { selectArticle } from 'store/slices/articleSlice';
-import { getCurrentTheme } from 'store/selectors';
+import { getCurrentTheme } from 'store/slices/theme';
 import SocialLinks from './SocialLinks';
 import ModalAuthor from './ModalAuthor';
 import defaultAvatar from 'images/admin/avatar_default.svg';
