+++ src/components/Intro/Intro.jsx
@@ -1,5 +1,5 @@
 // import { useSelector } from 'react-redux';
-// import { getCurrentTheme } from 'store/selectors';
+// import { getCurrentTheme } from 'store/slices/theme';
 import './Intro.css';
 import { useEffect, useRef } from 'react';
 import { useIsMobile } from 'utils/hooks/useIsMobile';
@@ -34,11 +34,9 @@ export default function Intro() {
   }, [slogan]);
 
   return (
-
-      <div className="intro__container">
-        <h1 className="intro__title">Design for all</h1>
-        <p className="intro__subtitle" ref={typeRef}></p>
-      </div>
-  
+    <div className="intro__container">
+      <h1 className="intro__title">Design for all</h1>
+      <p className="intro__subtitle" ref={typeRef}></p>
+    </div>
   );
 }
