+++ src/components/Main/SlideIframe/SlideIframe.jsx
@@ -1,6 +1,6 @@
 import { useRef, useMemo, useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
-import { getCurrentTheme } from 'store/selectors';
+import { getCurrentTheme } from 'store/slices/theme';
 import './SlideIframe.css';
 import { domain } from 'utils/config';
