+++ src/components/SearchInput/SearchInput.jsx
@@ -1,6 +1,6 @@
 import { useEffect, useRef, useState } from 'react';
 import { useSelector } from 'react-redux';
-import { getCurrentTheme } from 'store/selectors';
+import { getCurrentTheme } from 'store/slices/theme';
 
 import './SearchInput.css';
 import closeBtn from 'images/close-btn.svg';
