+++ src/store/index.js
@@ -1,7 +1,8 @@
 import { configureStore, combineReducers } from '@reduxjs/toolkit';
 import { persistReducer, persistStore } from 'redux-persist';
 import storage from 'redux-persist/lib/storage';
-import { themeReducer, userReducer, articleReducer } from './slices';
+import { userReducer, articleReducer } from './slices';
+import { themeReducer } from './slices/theme';
 
 const rootReducer = combineReducers({
   user: userReducer,
