+++ src/store/slices/theme/index.js
@@ -1,0 +1,3 @@
+export * from './middlewares';
+export * from './slice';
+export * from './selectors';
